+++ App.xcworkspace/contents.xcworkspacedata
@@ -10,4 +10,7 @@
    <FileRef
       location = "group:../Fair">
    </FileRef>
+   <FileRef
+      location = "group:../lottie-swift">
+   </FileRef>
 </Workspace>
