+++ Info.plist
@@ -46,11 +46,11 @@
 	<key>CFBundlePackageType</key>
 	<string>$(PRODUCT_BUNDLE_PACKAGE_TYPE)</string>
 	<key>CFBundleShortVersionString</key>
-	<string>0.0.18</string>
+	<string>0.0.19</string>
 	<key>CFBundleShortVersionString-FairComment</key>
 	<string>CFBundleShortVersionString should be a semantic version string that will match the tag name for the release of the app</string>
 	<key>CFBundleVersion</key>
-	<string>82</string>
+	<string>85</string>
 	<key>FairSymbol</key>
 	<string>hands.sparkles</string>
 	<key>FairSymbol-FairComment</key>
