+++ Package.swift
@@ -9,17 +9,12 @@ let package = Package(
     dependencies: [
         // the Fair main branch must be the first dependency to pass integration
         .package(name: "Fair", url: "https://fair-ground.org/Fair.git", .branch("main")),
-        // additional GitHub-hosted dependencies can be added below
-        //.package(url: "https://github.com/airbnb/lottie-ios.git", .branch("lottie/macos-spm")),
-        //.package(url: "https://github.com/airbnb/lottie-ios.git", .upToNextMajor(from: "3.2.3")),
-
-        // This fork (https://github.com/airbnb/lottie-ios/pull/1391) allows Lottie to be used from both macOS and iOS
-        .package(url: "https://github.com/msealand/lottie-ios.git", .branch("fix/multiplatform-spm")),
+        .package(url: "https://github.com/Lottie-Motion/lottie-swift.git", .branch("main")),
     ],
     targets: [
         .target(name: "App", dependencies: [
             .product(name: "FairApp", package: "Fair"),
-            .product(name: "Lottie", package: "lottie-ios"),
+            .product(name: "Lottie", package: "lottie-swift"),
         ], resources: [.process("Resources"), .copy("Bundle")]),
         .testTarget(name: "AppTests", dependencies: ["App"]),
     ]
