+++ Sources/App/Lottie-Motion.swift
@@ -1,0 +1,21 @@
+import FairApp
+
+// MARK: Parochial (package-local) Utilities
+
+/// Returns the localized string for the current module.
+///
+/// - Note: This is boilerplate package-local code that could be copied
+///  to any Swift package with localized strings.
+internal func loc(_ key: String, tableName: String? = nil, comment: String? = nil) -> String {
+    // TODO: use StringLocalizationKey
+    NSLocalizedString(key, tableName: tableName, bundle: .module, comment: comment ?? "")
+}
+
+/// Work-in-Progress marker
+@available(*, deprecated, message: "work in progress")
+internal func wip<T>(_ value: T) -> T { value }
+
+/// Intercept `LocalizedStringKey` constructor and forward it to ``SwiftUI.Text/init(_:bundle)``
+@usableFromInline internal func Text(_ string: LocalizedStringKey) -> SwiftUI.Text {
+    SwiftUI.Text(string, bundle: .module)
+}
